+++ tests/api/main.rs
@@ -1,3 +1,3 @@
 mod health_check;
 mod helpers;
-mod v1;
+mod api_v1;
